pyproject.toml
[project]
name = "agentoperator-outbound-workflow"
version = "0.1.0"
description = "A 7-skill outbound engine for Claude Code — ICP, targets, people, enrichment, research, writing, and deliverability-aware scheduling."
readme = "README.md"
license = {text = "MIT"}
requires-python = ">=3.10"
dependencies = [
    "requests>=2.31",
    "python-dotenv>=1.0",
    "pyyaml>=6.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4",
]

[build-system]
requires = ["setuptools>=68"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
include = ["clients*", "pipeline*", "config*"]
exclude = ["tests*", "examples*"]
